Добавил функционал на кнопку done, зарефакторил код, добавил module css, поправил оставшиеся todo
