+++ todolist/src/App.css
@@ -1,0 +1,20 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+#root {
+    display: flex;
+    justify-content: center;
+}
+
+.app {
+    width: 800px;
+}
+
+.header {
+    font-size: 40px;
+    font-weight: bold;
+    text-align: center;
+}
