+++ todolist/src/components/NoteCard/NoteCard.module.css
@@ -1,0 +1,22 @@
+.note {
+    display: flex;
+    padding: 15px;
+    border: 2px solid teal;
+    margin-top: 15px;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.noteColor {
+    display: flex;
+    padding: 15px;
+    border: 2px solid teal;
+    margin-top: 15px;
+    justify-content: space-between;
+    align-items: center;
+    background-color: aqua;
+}
+
+.description {
+    font-weight: bold;
+}
