+++ todolist/src/components/NoteForm/NoteForm.module.css
@@ -1,0 +1,5 @@
+.addBlock {
+    display: flex;
+    justify-content: space-between;
+    padding-top: 20px;
+}
