--- todolist/src/components/UI/button/MyButton.css
@@ -1,8 +1,0 @@
-.myBtn {
-    padding: 5px 15px;
-    color: teal;
-    font-size: 14px;
-    background: transparent;
-    border: 1px solid teal;
-    cursor: pointer;
-}
