+++ todolist/src/components/UI/button/MyButton.module.css
@@ -1,0 +1,8 @@
+.myBtn {
+    padding: 5px 15px;
+    color: teal;
+    font-size: 14px;
+    background: transparent;
+    border: 1px solid teal;
+    cursor: pointer;
+}
