--- todolist/src/components/UI/input/MyInput.css
@@ -1,6 +1,0 @@
-.myInput {
-    width: 70%;
-    padding: 5px 15px;
-    margin: 5px 0;
-    border: 1px solid teal;
-}
