+++ todolist/src/components/UI/input/MyInput.module.css
@@ -1,0 +1,6 @@
+.myInput {
+    width: 70%;
+    padding: 5px 15px;
+    margin: 5px 0;
+    border: 1px solid teal;
+}
