+++ todolist/src/components/UI/input/MyInput.tsx
@@ -1,21 +1,19 @@
-import './MyInput.css'
+import styles from './MyInput.module.css'
 
 type MyInputProps = {
     value?: string
     onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void
-    type?: string
     placeholder?: string
 }
 
 function MyInput ({
     value,
     onChange,
-    type,
     placeholder
 }: MyInputProps
 ) {
     return (
-        <input className='myInput' value={value} onChange={onChange} type={type} placeholder={placeholder}/>
+        <input className={styles.myInput} value={value} onChange={onChange} type='text' placeholder={placeholder}/>
     )
 }
