--- todolist/src/styles/App.css
@@ -1,29 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-#root {
-    display: flex;
-    justify-content: center;
-}
-
-.App {
-    width: 800px;
-}
-
-.note {
-    display: flex;
-    padding: 15px;
-    border: 2px solid teal;
-    margin-top: 15px;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.addBlock {
-    display: flex;
-    justify-content: space-between;
-    padding-top: 20px;
-}
